// wgg_front/src/lib/api/graphql_types/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Implement the DateTime<Utc> scalar
   *
   * The input/output is a string in RFC3339 format.
   */
  DateTime: any;
};
// ...
export type FreshLabel = {
  __typename?: 'FreshLabel';
  daysFresh: Scalars['Int'];
};
// ...
export type MoreButton = {
  __typename?: 'MoreButton';
  images: Array<Scalars['String']>;
};
// ...
export type NumberOfServings = {
  __typename?: 'NumberOfServings';
  amount: Scalars['Int'];
};
// ...
export type PrepTime = {
  __typename?: 'PrepTime';
  timeMinutes: Scalars['Int'];
};

export type ProductId = PromotionProduct & {
  __typename?: 'ProductId';
  id: Scalars['String'];
};

export type PromotionProduct = {
  id: Scalars['String'];
};

export enum Provider {
  Jumbo = 'JUMBO',
  Picnic = 'PICNIC'
}
// ...
export type SaleDescription = {
  __typename?: 'SaleDescription';
  text: Scalars['String'];
};

/**
 * Describes the type of sale that applies to the attached object.
 *
 * Think of "1 + 1 Free", or "50% off".
 */
export type SaleLabel = {
  __typename?: 'SaleLabel';
  text: Scalars['String'];
};

/** Until what date (inclusive) the attached sale is valid. */
export type SaleValidity = {
  __typename?: 'SaleValidity';
  validFrom: Scalars['DateTime'];
  validUntil: Scalars['DateTime'];
};

export type SubNutritionalItem = {
  __typename?: 'SubNutritionalItem';
  name: Scalars['String'];
  value: Scalars['String'];
};

/** If the item is unavailable */
export type UnavailableItem = {
  __typename?: 'UnavailableItem';
  explanationLong?: Maybe<Scalars['String']>;
  explanationShort?: Maybe<Scalars['String']>;
  reason: UnavailableReason;
  /**
   * Lists replacements if the store has suggested any.
   *
   * Some stores won't support this functionality, and this would therefore remain empty.
   */
  replacements: Array<WggSearchProduct>;
};
// ...
export type WggDecorator = FreshLabel | MoreButton | NumberOfServings | PrepTime | SaleDescription | SaleLabel | SaleValidity | UnavailableItem;
// ...
export type WggSaleCategory = {
  __typename?: 'WggSaleCategory';
  decorators: Array<WggDecorator>;
  id: Scalars['String'];
  imageUrls: Array<Scalars['String']>;
  /**
   * A potentially limited selection of items, only supported for certain [Provider]s.
   *
   * Picnic is one example of such a provider.
   * Generally recommended to query for more detailed information when needed.
   */
  limitedItems: Array<PromotionProduct>;
  name: Scalars['String'];
  provider: Provider;
};